Warning fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@26529 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi_rpc/socket.c
@@ -945,7 +945,7 @@ rpc_recvmsg(rcf_rpc_server *rpcs,
     snprintf(str_buf + strlen(str_buf), sizeof(str_buf) - strlen(str_buf),
              "), %s) -> %ld (%s)",
              send_recv_flags_rpc2str(flags),
-             out.retval, errno_rpc2str(RPC_ERRNO(rpcs)));
+             (long)out.retval, errno_rpc2str(RPC_ERRNO(rpcs)));
 
     TAPI_RPC_LOG("%s", str_buf);
